added WebcamInit class and moved lib instructions for Easy Open Cv to build
